Add useListingsApi to fetch ongoing/past listings
